added delete and pause project

+++ app/src/main/java/com/example/m/ismayilov/timetracker/adapter/OnlineAdapter.kt
@@ -72,7 +72,7 @@ class OnlineAdapter(val context: Context, var onClickLIstener: OnClickLIstener, 
             onClickLIstener.onClickListenerAction(holder.phone.text.toString())
         }
         holder.history.setOnClickListener {
-            onClickLIstener.onClickSetExpendValues(list.get(position).phone.toInt() , false)
+            onClickLIstener.onClickSetExpendValues(list.get(position).phone , false)
         }
     }
